refactoring of ast_annotations into core
