+++ src/awl/__init__.py
@@ -1,2 +1,1 @@
-from awl.ast_annotation import AstAnnotation  # noqa
-from awl.core import AstSerialization  # noqa
+from awl.core import ASTNotAModule, AstSerialization  # noqa
